Refactor main.php
Up until now, main.php was an all-in-one script which only served as a proof of concept. It was never meant to be final, so this commit sets us down a better direction.

* main.php (to be renamed) is now the bare minimum code to pull all the functionality together.
* Book.php: a class to represent a book's metadata
* BookBuilder.php: a class to build a book from user input.
* CommandBuilder.php: a class to build the minecraft command from a Book object.

Note there is no functional change from the last commit. This is strictly a code reorganization. All the functionality and lack thereof, including bugs, is still here.