Add book-building functionality to Book
Before, the Book class was fairly simple: just attributes that make up a book with simple setters and getters.

Now, the Book class has functionality. A new build() method was added to convert the basic attributes into a book. This is done with two key private methods, parse() and paginate().

* parse() is responsible for taking the initial, raw, supplied text and breaking it into a clean structure: an array of lines (or paragraphs) and each line/paragraph is an array of words
* paginate() loops through the parsed structure and mimic's the Minecraft in-game line and page breaks